+++ extras/tests/JsonVariant/as.cpp
@@ -136,7 +136,7 @@ TEST_CASE("JsonVariant::as()") {
 
     REQUIRE(variant.as<long>() == 42L);
     REQUIRE(variant.as<JsonString>() == "42");
-    REQUIRE(variant.as<JsonString>().isStatic() == true);
+    REQUIRE(variant.as<JsonString>().isLinked() == true);
   }
 
   SECTION("set(\"hello\")") {
@@ -159,7 +159,7 @@ TEST_CASE("JsonVariant::as()") {
     REQUIRE(variant.as<const char*>() == std::string("4.2"));
     REQUIRE(variant.as<std::string>() == std::string("4.2"));
     REQUIRE(variant.as<JsonString>() == "4.2");
-    REQUIRE(variant.as<JsonString>().isStatic() == false);
+    REQUIRE(variant.as<JsonString>().isLinked() == false);
   }
 
   SECTION("set(\"true\")") {