+++ src/ArduinoJson/Collection/CollectionImpl.hpp
@@ -64,7 +64,7 @@ inline bool CollectionData::copyFrom(const CollectionData& src,
   for (VariantSlot* s = src._head; s; s = s->next()) {
     VariantData* var;
     if (s->key() != 0) {
-      String key(s->key(), !s->ownsKey());
+      String key(s->key(), s->ownsKey() ? String::Copied : String::Linked);
       var = addMember(adaptString(key), pool, getStringStoragePolicy(key));
     } else {
       var = addElement(pool);