+++ src/ArduinoJson/Strings/String.hpp
@@ -14,15 +14,15 @@ namespace ARDUINOJSON_NAMESPACE {
 
 class String : public SafeBoolIdom<String> {
  public:
-  String() : _data(0), _size(0), _isStatic(true) {}
+  enum Ownership { Copied, Linked };
 
-  String(const char* data, bool isStaticData = true)
-      : _data(data),
-        _size(data ? ::strlen(data) : 0),
-        _isStatic(isStaticData) {}
+  String() : _data(0), _size(0), _ownership(Linked) {}
 
-  String(const char* data, size_t sz, bool isStaticData = true)
-      : _data(data), _size(sz), _isStatic(isStaticData) {}
+  String(const char* data, Ownership ownership = Linked)
+      : _data(data), _size(data ? ::strlen(data) : 0), _ownership(ownership) {}
+
+  String(const char* data, size_t sz, Ownership ownership = Linked)
+      : _data(data), _size(sz), _ownership(ownership) {}
 
   const char* c_str() const {
     return _data;
@@ -32,8 +32,8 @@ class String : public SafeBoolIdom<String> {
     return !_data;
   }
 
-  bool isStatic() const {
-    return _isStatic;
+  bool isLinked() const {
+    return _ownership == Linked;
   }
 
   size_t size() const {
@@ -75,7 +75,7 @@ class String : public SafeBoolIdom<String> {
  private:
   const char* _data;
   size_t _size;
-  bool _isStatic;
+  Ownership _ownership;
 };
 
 }  // namespace ARDUINOJSON_NAMESPACE